Sound struct added
Sound struct added in Model of MVC successfully . The playSound func has been moved to Sound struct
