+++ Quiz App/Model/QuestionBrain.swift
@@ -10,6 +10,7 @@ import Foundation
 struct QuestionBrain {
     
     var questionNumber = 0
+    var progress: Float = 1
     
     let quiz = [
         Questions(ask: "Four + Two is equal to Six", reply: "False"),
@@ -33,6 +34,20 @@ struct QuestionBrain {
        return quiz[questionNumber].ask
     }
     
+    mutating func nextQuestion() {
+        // it's loop for count up question in quiz array and countdown progress bar
+        if questionNumber + 1 < quiz.count {
+            questionNumber += 1
+            progress -= 1 / Float(quiz.count)
+
+        } else {
+            questionNumber = 0
+            progress = 1
+
+        }
+        print("QB: questionNumber \(questionNumber)")
+        print("QB: progress \(progress)")
+    }
  
     
 }
